Release 1.0.0-beta07 (fixes #128).

+++ sources/engine/SiliconStudio.Paradox.Engine/Effects/IModelInstance.cs
@@ -24,6 +24,6 @@ namespace SiliconStudio.Paradox.Effects
         /// Gets the draw order.
         /// </summary>
         /// <value>The draw order.</value>
-        float DrawOrder { get; }
+        int DrawOrder { get; }
     }
 }
